+++ src/dve/core_engine/type_hints.py
@@ -244,3 +244,12 @@ SUBMISSION_RESULTS: tuple[SubmissionResult, ...] = tuple(list(get_args(Submissio
 
 BinaryComparator = Callable[[Any, Any], bool]
 """Type hint for operator functions"""
+
+DVEStageName = Literal[
+    "audit_received",
+    "file_transformation",
+    "data_contract",
+    "business_rules",
+    "error_report",
+    "pipeline",
+]